requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.32.1
sqlalchemy==2.0.36
psycopg2-binary==2.9.10
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0
pydantic==2.10.3
pydantic-settings==2.7.0
python-dotenv==1.0.1
python-multipart==0.0.18
jinja2==3.1.4
email-validator==2.2.0
